+++ CHANGELOG.md
@@ -38,7 +38,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## Unreleased
 
 ### Bug Fixes
+
 * [#138](https://github.com/babylonlabs-io/vigilante/pull/138) fix: panic in SendCheckpointToBTC
+
+### Improvements
+
+* [#136](https://github.com/babylonlabs-io/vigilante/pull/136) rate limit activations
 
 ## v0.18.0